@ggerganov 2025-12-19
I'm not yet sure it is worth adding this additional implementation. Can it not be consolidated with the rest of the repacking schemes?

cc @Alcpz for opinions