@yuanjia111 2026-01-07
> @yuanjia11我离开了一段时间，我会看看的。

Thank you, @Alcpz! 😊  
I completely understand you’ve been away — please take your time.  
I just wanted to gently follow up since your input on this design direction would mean a lot.  
Looking forward to your feedback whenever you’re ready!