@elfarolab 2026-01-21
@hongyang-7 

if you like I would like to help testing this PR.

I can build on nVidia Jetson AGX Orin and I am interested to improvements related to Q4_K  quants.

Please could you tell me the building options if any specific? 

Also I have very limited free RAM memory (~ 16 GB) due the embedded Linux env. 
What model would you like me to test?

Thank you so much.